requirements.txt
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
langchain-core>=0.3.17
python-dotenv
streamlit
validators
youtube-transcript-api
pytube
unstructured
beautifulsoup4
requests
pypdf
groq
duckduckgo-search
gTTS
pandas
numexpr
fastapi
uvicorn
httpx
python-multipart
python-dotenv
